Made musicField and focusField radial spinners. Added ability to overwrite NumberFormat for NumberTextField.

+++ HomeworkData/src/CustomControls/RadialSpinnerSkin.java
@@ -82,7 +82,7 @@ public class RadialSpinnerSkin extends BehaviorSkinBase<RadialSpinner, RadialSpi
 		numField.textProperty().addListener((observable, oldValue, newValue) -> {
 			if (!numField.getText().equals("")) {
 				numField.parseAndFormatInput();
-				control.setValue(numField.getNumber().intValue());
+				control.setValue(numField.getNumber().doubleValue());
 				updatePosition(control);
 			}
 		});
